Make period_start_date editable
The field `period_start_date` on ApiLimits was set to `auto_add_now`.
This makes the field read-only, and hides it in the Django admin. In
order to make the field editable in the admin, it has been updated to
use `timezome.now` as a default and forbid null values instead of using
`auto_add_now`.